feat: integrate TopPanel and AddMoreCoursesButton into GraphView. TopPanel makes the Panel UI components responsive.
