+++ client/src/components/SearchComponent.jsx
@@ -1,10 +1,5 @@
-import { Panel } from "@xyflow/react";
 import { NodeSearch } from "./node-search";
 
 export function SearchComponent() {
-  return (
-    <Panel position="top-center">
-      <NodeSearch />
-    </Panel>
-  );
+  return <NodeSearch />;
 }
